Added Level 01, Made Character Changes
The character is now using the Standard Assets Rigidbody First Person Controller, modified to better suit the Fysics Character for desired physics interactions.

Also, a simple level 01 has been added, with object dispensers, a small playing field and the beginnings of a network host (still using HLAPI, which is deprecated - could not find documentation for other solutions).

The character should be modified further to have an animated body, giving both players better feedback as to their position and look direction.