Rope & Unity Physics
Created a rope which was experiencing problems with the current Physics System. After running some tests, it would seem these issues are removed with the new Unity Physics System (uses ECS still in its preview phase) (uses Unity Physics still in its experimental phase) (still needs to be fully tested if this fixes the rope physics or not).

Furthermore, the new Physics System is a more robust and accurate physics simulation, as it can precalculate trajectories inside a single frame of bouncing against objects.